Add Car and Search Car
This commit includes the implementation of the Add Cars and Search Cars
features. The Search Cars feature includes searching for a car by VIN
number, searching for all car with a given make, model, and year,
searching for all cars with a sale price less than a given price and
searching for all the features of a car with a given VIN number.